Implementation of boto3 instead of rest api.
