+++ README.md
@@ -67,6 +67,7 @@ Config saved [/Users/allexx/.aps_config]
 
 ```
 apsconnect install-backend --name NAME --image IMAGE --config-file CONFIG_FILE --hostname HOSTNAME \
+                          [--aws-ecr-key AWS_ECR_KEY] [--aws-ecr-secret AWS_ECR_SECRET] \
                           [--healthcheck-path HEALTHCHECK_PATH] [--root-path ROOT_PATH] \
                           [--namespace NAMESPACE] [--replicas REPLICAS] [--tls-secret-name TLS_SECRET_NAME] \
                           [--force FORCE]
