+++ apsconnectcli/awsmanager/ecr.py
@@ -1,9 +1,25 @@
-import json
+import boto3
 import logging
 
-from apsconnectcli.awsmanager.aws import AWSClient
-
 logger = logging.getLogger(__name__)
+
+
+class AWSAPIException(Exception):
+    def __init___(self, error_text):
+        Exception.__init__(self, "Error while calling AWS API. Error details from AWS: {0}".
+                           format(error_text))
+
+
+class AWSClient(object):
+    """Parent client class for asw services"""
+    def __init__(self, service_name, region, key_id, access_key):
+        session = boto3.session.Session()
+        self.client = session.client(
+            service_name=service_name,
+            region_name=region,
+            aws_access_key_id=key_id,
+            aws_secret_access_key=access_key,
+        )
 
 
 class ECRClient(AWSClient):
@@ -12,15 +28,8 @@ class ECRClient(AWSClient):
 
     def __init__(self, region, key_id, access_key):
         super(ECRClient, self) \
-            .__init__(ECRClient.serviceName, region, key_id, access_key)
+            .__init__(self.serviceName, region, key_id, access_key)
 
-    def get_authorization_token(self, registry_ids=None):
-        url = self.resolve_api_url()
-
-        if not registry_ids:
-            params = '{}'
-        else:
-            params = json.dumps({"registryIds": registry_ids})
-        aws_target = 'AmazonEC2ContainerRegistry_V20150921.GetAuthorizationToken'
-
-        return self.call_api(url, 'POST', params, aws_target)
+    def get_auth_token(self, registry_id=None):
+        response = self.client.get_authorization_token(registryIds=[registry_id, ])
+        return response
